Refactor build-all.sh script to remove vite.config.ts file if it exists

+++ pipeline/build-all.sh
@@ -182,7 +182,9 @@ import { AppInfoClz } from \"@/app/__CORE__/meta/ctypes\"
             rm -f *lock*
             [ ! -d node_modules ] && npm install --omit=dev --force 
             # [ ! -d node_modules ] && npm install -S -D --force 
-            rm ./vite.config.ts
+            if [ -e ./vite.config.ts ]; then
+                rm ./vite.config.ts
+            fi
             rm -rf .next
             npm run build
             cd .next
